# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "keeper project",
        "title": "Keeper Server",
        "contact": {
            "name": "Nexbug",
            "url": "https://github.com/nextlag",
            "email": "nextbug@ya.ru"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/notes": {
            "get": {
                "description": "Retrieve all notes for the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Get all notes for the current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.SecretNote"
                            }
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            },
            "post": {
                "description": "Upload a new note for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Add a new note",
                "parameters": [
                    {
                        "description": "Note data",
                        "name": "note",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.SecretNote"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/entity.SecretNote"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/user/notes/{id}": {
            "delete": {
                "description": "Delete a specific note identified by its UUID",
                "tags": [
                    "notes"
                ],
                "summary": "Delete a note by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Delete accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a specific note identified by its UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Update a note by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated note data",
                        "name": "note",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.SecretNote"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Update accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entity.Meta": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "Name or type of the metadata.",
                    "type": "string"
                },
                "value": {
                    "description": "Value of the metadata.",
                    "type": "string"
                }
            }
        },
        "entity.SecretNote": {
            "type": "object",
            "properties": {
                "meta": {
                    "description": "Associated metadata for the note.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Meta"
                    }
                },
                "name": {
                    "description": "Name or title of the note.",
                    "type": "string"
                },
                "note": {
                    "description": "Content of the note.",
                    "type": "string"
                }
            }
        },
...
        "v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                },
                "status": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}